Extract reusable UI components into separate modules
