+++ core/designsystem/src/main/kotlin/com/oussamateyib/thoth/core/designsystem/theme/NoteColors.kt
@@ -1,0 +1,5 @@
+package com.oussamateyib.thoth.core.designsystem.theme
+
+object NoteColors {
+    val noteColors = listOf(RedPink, RedOrange, LightGreen, Violet, BabyBlue, White)
+}
