+++ core/ui/build.gradle.kts
@@ -11,7 +11,7 @@ java {
 }
 
 android {
-    namespace = "com.oussamateyib.thoth.core.designsystem"
+    namespace = "com.oussamateyib.thoth.core.ui"
     compileSdk = 37
 
     defaultConfig {
@@ -25,14 +25,20 @@ android {
 }
 
 dependencies {
-    implementation(platform(libs.androidx.compose.bom))
+    implementation(projects.core.designsystem)
+    api(projects.core.domain)
+    api(projects.core.model)
 
+    implementation(platform(libs.androidx.compose.bom))
+    implementation(libs.androidx.core)
     implementation(libs.androidx.compose.foundation)
     implementation(libs.androidx.compose.foundation.layout)
-    api(libs.androidx.compose.material3)
+    implementation(libs.androidx.compose.material3)
+    api(libs.androidx.compose.runtime)
     api(libs.androidx.compose.ui)
+    implementation(libs.androidx.compose.ui.geometry)
     api(libs.androidx.compose.ui.graphics)
     api(libs.androidx.compose.ui.text)
     implementation(libs.androidx.compose.ui.unit)
-    api(libs.androidx.compose.runtime)
+    implementation(libs.colorpicker.compose)
 }
