+++ core/ui/src/main/kotlin/com/oussamateyib/thoth/core/ui/NoteOrderSection.kt
@@ -1,4 +1,4 @@
-package com.oussamateyib.thoth.feature.notes.impl.list.components
+package com.oussamateyib.thoth.core.ui
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -9,12 +9,12 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
+import com.oussamateyib.thoth.core.designsystem.components.LabeledRadioButton
 import com.oussamateyib.thoth.core.domain.util.NoteOrder
 import com.oussamateyib.thoth.core.domain.util.OrderType
-import com.oussamateyib.thoth.feature.notes.impl.R
 
 @Composable
-internal fun OrderSection(
+fun NoteOrderSection(
     modifier: Modifier = Modifier,
     noteOrder: NoteOrder = NoteOrder.Date(OrderType.Descending),
     onOrderChange: (NoteOrder) -> Unit
@@ -25,19 +25,19 @@ internal fun OrderSection(
         Row(
             modifier = Modifier.fillMaxWidth()
         ) {
-            DefaultRadioButton(
+            LabeledRadioButton(
                 text = stringResource(R.string.color),
                 selected = noteOrder is NoteOrder.Color,
                 onSelect = { onOrderChange(NoteOrder.Color(noteOrder.orderType)) }
             )
             Spacer(modifier = Modifier.width(8.dp))
-            DefaultRadioButton(
+            LabeledRadioButton(
                 text = stringResource(R.string.date),
                 selected = noteOrder is NoteOrder.Date,
                 onSelect = { onOrderChange(NoteOrder.Date(noteOrder.orderType)) }
             )
             Spacer(modifier = Modifier.width(8.dp))
-            DefaultRadioButton(
+            LabeledRadioButton(
                 text = stringResource(R.string.title),
                 selected = noteOrder is NoteOrder.Title,
                 onSelect = { onOrderChange(NoteOrder.Title(noteOrder.orderType)) }
@@ -46,13 +46,13 @@ internal fun OrderSection(
         Row(
             modifier = Modifier.fillMaxWidth()
         ) {
-            DefaultRadioButton(
+            LabeledRadioButton(
                 text = stringResource(R.string.ascending),
                 selected = noteOrder.orderType is OrderType.Ascending,
                 onSelect = { onOrderChange(noteOrder.copy(OrderType.Ascending)) }
             )
             Spacer(modifier = Modifier.width(8.dp))
-            DefaultRadioButton(
+            LabeledRadioButton(
                 text = stringResource(R.string.descending),
                 selected = noteOrder.orderType is OrderType.Descending,
                 onSelect = { onOrderChange(noteOrder.copy(OrderType.Descending)) }
