+++ core/ui/src/main/res/drawable/delete.xml
