+++ feature/notes/impl/build.gradle.kts
@@ -31,11 +31,11 @@ dependencies {
     api(projects.core.domain)
     api(projects.core.model)
     api(projects.core.navigation)
+    implementation(projects.core.ui)
     implementation(projects.feature.notes.api)
 
     implementation(platform(libs.androidx.compose.bom))
     implementation(libs.androidx.annotation)
-    implementation(libs.androidx.core)
     implementation(libs.androidx.compose.animation)
     implementation(libs.androidx.compose.animation.core)
     implementation(libs.androidx.compose.foundation)
@@ -43,7 +43,6 @@ dependencies {
     implementation(libs.androidx.compose.material3)
     api(libs.androidx.compose.runtime)
     api(libs.androidx.compose.ui)
-    implementation(libs.androidx.compose.ui.geometry)
     api(libs.androidx.compose.ui.graphics)
     api(libs.androidx.compose.ui.text)
     implementation(libs.androidx.compose.ui.unit)
@@ -52,7 +51,6 @@ dependencies {
     api(libs.androidx.lifecycle.viewmodel)
     implementation(libs.androidx.lifecycle.viewmodel.compose)
     api(libs.androidx.navigation3.runtime)
-    implementation(libs.colorpicker.compose)
     api(libs.dagger)
     implementation(libs.dagger.hilt.core)
     implementation(libs.hilt.android)
