--- feature/notes/impl/src/main/kotlin/com/oussamateyib/thoth/feature/notes/impl/editor/util/NoteColors.kt
@@ -1,12 +1,0 @@
-package com.oussamateyib.thoth.feature.notes.impl.editor.util
-
-import com.oussamateyib.thoth.core.designsystem.theme.BabyBlue
-import com.oussamateyib.thoth.core.designsystem.theme.LightGreen
-import com.oussamateyib.thoth.core.designsystem.theme.RedOrange
-import com.oussamateyib.thoth.core.designsystem.theme.RedPink
-import com.oussamateyib.thoth.core.designsystem.theme.Violet
-import com.oussamateyib.thoth.core.designsystem.theme.White
-
-object NoteColors {
-    val noteColors = listOf(RedPink, RedOrange, LightGreen, Violet, BabyBlue, White)
-}
