+++ feature/notes/impl/src/main/res/values/strings.xml
@@ -1,19 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <string name="add_note">Add note</string>
-    <string name="ascending">Ascending</string>
     <string name="back">Back</string>
-    <string name="color">Color</string>
-    <string name="date">Date</string>
-    <string name="delete_note">Delete note</string>
-    <string name="descending">Descending</string>
     <string name="pick_color">Pick color</string>
     <string name="list_screen_top_bar_title">My Notes</string>
     <string name="note_content_hint">What’s on your mind…</string>
     <string name="note_deleted">Note deleted</string>
     <string name="note_not_found">This note could not be found.</string>
     <string name="note_title_hint">Give it a title…</string>
-    <string name="title">Title</string>
     <string name="sort_notes">Sort notes</string>
     <string name="undo">Undo</string>
 </resources>
