+++ settings.gradle.kts
@@ -41,5 +41,6 @@ include(":core:domain")
 include(":core:designsystem")
 include(":core:model")
 include(":core:navigation")
+include(":core:ui")
 include(":feature:notes:api")
 include(":feature:notes:impl")
